# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing group food orders.
  title: GroupGrub
  version: 1.0.0
servers:
  - description: Production server
    url: https://api.groupgrub.com
paths:
  /groups:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Group'
                type: array
          description: Successfully retrieved a list of all groups in the service.
        "404":
          description: No groups found in the service.
      summary: Retrieve a list of all groups in the service.
      x-accepts:
        - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        description: "Group object to add, must specify group name, participant IDs,\
          \ groupOrder IDs, and administrator ID."
        required: true
      responses:
        "201":
          description: Successfully added a new group.
        "400":
          description: Failed to add a new group.
      summary: Add a new group to the service.
      x-content-type: application/json
      x-accepts:
        - application/json
  /groups/{groupId}:
    delete:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "204":
          description: Group deleted successfully.
        "404":
          description: Group not found.
      summary: Delete a group.
      x-accepts:
        - application/json
    get:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
          description: Group details retrieved successfully.
        "404":
          description: Group not found.
      summary: Retrieve details of a specific group.
      x-accepts:
        - application/json
    put:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        "200":
          description: Group updated successfully.
        "404":
          description: Group not found.
      summary: Update a group's details.
      x-content-type: application/json
      x-accepts:
        - application/json
  /groups/{groupId}/orders:
    get:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GroupOrderResponse'
                type: array
          description: List of group orders retrieved successfully.
      summary: Retrieve all group orders for a specific group.
      x-accepts:
        - application/json
    post:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupOrder'
        required: true
      responses:
        "201":
          description: Group order created successfully.
      summary: Create a new group order.
      x-content-type: application/json
      x-accepts:
        - application/json
  /groups/{groupId}/orders/{orderId}:
    delete:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
        - description: The unique identifier of the order.
          explode: false
          in: path
          name: orderId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "204":
          description: Group order deleted successfully.
        "404":
          description: Group order not found.
      summary: Delete a group order.
      x-accepts:
        - application/json
    get:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
        - description: The unique identifier of the order.
          explode: false
          in: path
          name: orderId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupOrderResponse'
          description: Group order details retrieved successfully.
        "404":
          description: Group order not found.
      summary: Retrieve details of a specific group order.
      x-accepts:
        - application/json
    put:
      parameters:
        - description: The unique identifier of the group.
          explode: false
          in: path
          name: groupId
          required: true
          schema:
            type: string
          style: simple
        - description: The unique identifier of the order.
          explode: false
          in: path
          name: orderId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupOrder'
        required: true
      responses:
        "200":
          description: Group order updated successfully.
        "400":
          description: Failed to update the group order.
      summary: Update a group order.
      x-content-type: application/json
      x-accepts:
        - application/json
...
components:
  schemas:
    Group:
      example:
        administratorID: administratorID
        groupOrderIDs:
          - groupOrderIDs
          - groupOrderIDs
        groupID: groupID
        name: name
        participantIDs:
          - participantIDs
          - participantIDs
      properties:
        groupID:
          type: string
        name:
          type: string
        participantIDs:
          items:
            type: string
          type: array
        groupOrderIDs:
          items:
            type: string
          type: array
        administratorID:
          type: string
      type: object
    GroupOrder:
      example:
        participantOrderIDs:
          - participantOrderIDs
          - participantOrderIDs
        groupOrderID: groupOrderID
        menuItemIDs:
          key: 0
        desiredPickupTimeframe: desiredPickupTimeframe
        foodProviderID: foodProviderID
        status: status
      properties:
        groupOrderID:
          type: string
        status:
          type: string
        menuItemIDs:
          additionalProperties:
            type: integer
          type: object
        participantOrderIDs:
          items:
            type: string
          type: array
        desiredPickupTimeframe:
          type: string
        foodProviderID:
          type: string
      type: object
    GroupOrderResponse:
      allOf:
        - $ref: '#/components/schemas/GroupOrder'
        - properties:
            participantOrders:
              items:
                $ref: '#/components/schemas/ParticipantOrder'
              type: array
          type: object
      example:
        participantOrderIDs:
          - participantOrderIDs
          - participantOrderIDs
        groupOrderID: groupOrderID
        menuItemIDs:
          key: 0
        participantOrders:
          - menuItemIDs:
              key: 6
            comments: comments
            participantOrderID: participantOrderID
          - menuItemIDs:
              key: 6
            comments: comments
            participantOrderID: participantOrderID
        desiredPickupTimeframe: desiredPickupTimeframe
        foodProviderID: foodProviderID
        status: status
...
    ParticipantOrder:
      example:
        menuItemIDs:
          key: 6
        comments: comments
        participantOrderID: participantOrderID
      properties:
        participantOrderID:
          type: string
        menuItemIDs:
          additionalProperties:
            type: integer
          type: object
        comments:
          type: string
      type: object
...